Updated alpha-diversity and some function names
The function pd from the picante package was giving some weird errors during subsetting because it was generating unrooted subtrees. So I re-wrote the function to mid-point root any subtree that was unrooted. This required adding the phytools package as a dependency, but I remove picante, as pd was the only function from that package. I also had to rename some functions (chaning . to _) because it was causing issues with documenting the package with roxygen2